Refactor: shuffle関数をutilsに移動
