+++ app/features/slide/components/WordsSlide.tsx
@@ -38,6 +38,7 @@ import {
 	TooltipTriggerNoButton,
 } from "~/components/ui/tooltip";
 import type { SlideWord } from "../types";
+import { shuffleSlides } from "../utils/shuffle";
 
 interface WordsSlideProps {
 	data: SlideWord[];
@@ -51,10 +52,9 @@ const WordsSlide = ({ data, isSizing, setIsSizing }: WordsSlideProps) => {
 		playbackSpeed: 5,
 		isLooping: true,
 		showImages: true,
-		data: data, // data を state に保持
+		data: data,
 	});
 
-	// キーバインドの処理
 	const handleKeyDown = useCallback(
 		(event: KeyboardEvent) => {
 			switch (event.key) {
@@ -85,19 +85,6 @@ const WordsSlide = ({ data, isSizing, setIsSizing }: WordsSlideProps) => {
 		};
 	}, [handleKeyDown]);
 
-	// シャッフル関数
-	const shuffleSlides = (data: SlideWord[]) => {
-		const grouped = data.reduce<Record<string, SlideWord[]>>((acc, item) => {
-			const pairId = item.id.split("-")[0];
-			if (!acc[pairId]) acc[pairId] = [];
-			acc[pairId].push(item);
-			return acc;
-		}, {});
-		return Object.values(grouped)
-			.sort(() => Math.random() - 0.5)
-			.flat();
-	};
-
 	const handleShuffle = () => {
 		setSlide((prev) => ({
 			...prev,
@@ -119,7 +106,7 @@ const WordsSlide = ({ data, isSizing, setIsSizing }: WordsSlideProps) => {
 		</>
 	) : (
 		<div className="flex flex-col justify-center items-center">
-			<div className={isSizing ? "w-10/12" : "w-8/12"}>
+			<div className={isSizing ? "w-10/12" : "w-9/12"}>
 				<Carousel
 					opts={{ loop: slide.isLooping }}
 					plugins={plugins}
@@ -133,36 +120,37 @@ const WordsSlide = ({ data, isSizing, setIsSizing }: WordsSlideProps) => {
 								key={item.id}
 								aria-label={`スライド: ${item.word || item.translation}`}
 							>
-								<div className="mt-2 mb-3 p-1">
+								<div className="my-3 p-1">
 									<Card className="rounded-md">
 										<CardContent className="flex flex-col items-center aspect-square justify-center p-6">
-											{/* 単語データ */}
 											{item.word && (
 												<div className="relative w-full h-full flex items-center justify-center">
 													<p className="text-2xl sm:text-4xl font-semibold text-center">
 														{item.word}
 													</p>
-													<div className="absolute bottom-2 left-2">
-														<WordDetails
-															word={item.word}
-															triggerElement={
-																<Button variant="ghost" size="icon">
-																	<Info />
-																</Button>
-															}
-														/>
-													</div>
-													<p className="absolute bottom-2 left-1/2 transform -translate-x-1/2">
-														{Math.ceil((slide.data.indexOf(item) + 1) / 2)}/
-														{Math.ceil(data.length / 2)}
-													</p>
-													<div className="absolute bottom-2 right-2">
-														<Speech word={item.word} />
+													<div className="absolute bottom-0 left-0 right-0 flex justify-between items-center px-2">
+														<div>
+															<WordDetails
+																word={item.word}
+																triggerElement={
+																	<Button variant="ghost" size="icon">
+																		<Info />
+																	</Button>
+																}
+															/>
+														</div>
+
+														<p className="text-center">
+															{Math.ceil((slide.data.indexOf(item) + 1) / 2)}/
+															{Math.ceil(data.length / 2)}
+														</p>
+
+														<div>
+															<Speech word={item.word} />
+														</div>
 													</div>
 												</div>
 											)}
-
-											{/* 翻訳データ */}
 											{item.translation && (
 												<div className="relative flex items-center justify-center w-full h-full">
 													{/* 背景画像 */}
@@ -189,8 +177,7 @@ const WordsSlide = ({ data, isSizing, setIsSizing }: WordsSlideProps) => {
 					<CarouselNext />
 				</Carousel>
 
-				<div className="flex justify-center space-x-3">
-					{/* ▶ / ⏸ 再生・停止 */}
+				<div className="mb-4 flex justify-center space-x-3">
 					<TooltipProvider>
 						<Tooltip>
 							<TooltipTriggerNoButton>
@@ -223,7 +210,6 @@ const WordsSlide = ({ data, isSizing, setIsSizing }: WordsSlideProps) => {
 						</Tooltip>
 					</TooltipProvider>
 
-					{/* ⏳ スライド1枚あたりの表示時間設定 */}
 					<TooltipProvider>
 						<Tooltip>
 							<TooltipTriggerNoButton>
@@ -263,7 +249,6 @@ const WordsSlide = ({ data, isSizing, setIsSizing }: WordsSlideProps) => {
 						</Tooltip>
 					</TooltipProvider>
 
-					{/* 🖼 / 🚫🖼 画像表示・非表示 */}
 					<TooltipProvider>
 						<Tooltip>
 							<TooltipTriggerNoButton>
@@ -286,7 +271,6 @@ const WordsSlide = ({ data, isSizing, setIsSizing }: WordsSlideProps) => {
 						</Tooltip>
 					</TooltipProvider>
 
-					{/* ♾ ループ切り替え */}
 					<TooltipProvider>
 						<Tooltip>
 							<TooltipTriggerNoButton>
@@ -323,7 +307,6 @@ const WordsSlide = ({ data, isSizing, setIsSizing }: WordsSlideProps) => {
 						</Tooltip>
 					</TooltipProvider>
 
-					{/* ⛶ / ⛌ 拡大・縮小 */}
 					<TooltipProvider>
 						<Tooltip>
 							<TooltipTriggerNoButton>
